requirements.txt
flask>=2.3.0
gunicorn>=21.2.0
google-cloud-aiplatform>=1.71.1
google-cloud-secret-manager>=2.16.0
google-genai>=0.8.0
weasyprint>=60.0
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
matplotlib>=3.8.0
openpyxl>=3.1.0
